benchmark/requirements.txt
dnspython
tqdm
